Debug and fix add to cart backend and database features
