+++ src/main/java/com/codecool/recipeme/model/ShoppingCart.java
@@ -1,27 +1,25 @@
 package com.codecool.recipeme.model;
 
 import com.codecool.recipeme.model.generated.Recipe;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import java.util.List;
 
 @Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Builder
 public class ShoppingCart {
     @Id
     @GeneratedValue
-    long id;
+    Long id;
 
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "shoppingCart")
-    @ElementCollection
+    @OneToMany(mappedBy = "shoppingCart", cascade = CascadeType.ALL)
     List<Recipe> recipes;
-
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
 
 }
