+++ src/main/java/com/codecool/recipeme/repository/ShoppingCartRepository.java
@@ -1,10 +1,10 @@
 package com.codecool.recipeme.repository;
 
-import com.codecool.recipeme.model.generated.Recipe;
+import com.codecool.recipeme.model.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 
 @Component
-public interface ShoppingCartRepository extends JpaRepository<Recipe, Long> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
 }
