+++ src/main/java/com/codecool/recipeme/service/RecipeSearchService.java
@@ -4,7 +4,7 @@ import com.codecool.recipeme.model.Search;
 import com.codecool.recipeme.model.generated.HitsItem;
 import com.codecool.recipeme.model.generated.Recipe;
 import com.codecool.recipeme.model.generated.Response;
-import com.codecool.recipeme.repository.RecipeRepository;
+import com.codecool.recipeme.repository.RecipeApiRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -15,10 +15,10 @@ import java.util.stream.Collectors;
 public class RecipeSearchService {
 
     @Autowired
-    RecipeRepository recipeRepository;
+    RecipeApiRepository recipeApiRepository;
 
     public List<Recipe> getRecipesBySearch(Search search) throws IllegalAccessException, NoSuchFieldException {
-        Response response = recipeRepository.getResponse(search);
+        Response response = recipeApiRepository.getResponse(search);
         return response.getHits().stream().map(HitsItem::getRecipe).collect(Collectors.toList());
     }
 }
