+++ src/main/java/com/codecool/recipeme/service/ShoppingCartService.java
@@ -1,6 +1,8 @@
 package com.codecool.recipeme.service;
 
+import com.codecool.recipeme.model.ShoppingCart;
 import com.codecool.recipeme.model.generated.Recipe;
+import com.codecool.recipeme.repository.RecipeRepository;
 import com.codecool.recipeme.repository.ShoppingCartRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -13,11 +15,16 @@ public class ShoppingCartService {
     @Autowired
     ShoppingCartRepository shoppingCartRepository;
 
+    @Autowired
+    RecipeRepository recipeRepository;
+
     public List<Recipe> getRecipesFromShoppingCart() {
-        return shoppingCartRepository.findAll();
+        return recipeRepository.findAllByShoppingCartId(1L);
     }
 
     public void addRecipesToShoppingCart(Recipe recipe) {
-        shoppingCartRepository.save(recipe);
+        ShoppingCart shoppingCart = shoppingCartRepository.getOne(1L);
+        recipe.setShoppingCart(shoppingCart);
+        recipeRepository.saveAndFlush(recipe);
     }
 }
